Add spamoor to ansible collection (#318)
I've noticed that [spamoor](https://github.com/ethpandaops/spamoor) was
missing from the ansible collection, so I've added it to the collection.

You can use the following flags to modify the basic behavior of spamoor
```yaml
spamoor_rpc_urls:
  - your-execution-node:8545
spamoor_rpc_throughput: "10"
spamoor_scenario: "eoatx"
spamoor_private_key: ""
```
and also override `spamoor_container_command` to modify other things

Please let me know if there's anything that needs fixing! 😄 
If this can be merged, I'd also like to work on a follow-up pr on adding
this role to the devnet templates so that we can use this in devnets in
lieu of tx-fuzz.