fix(teku): Adjust teku so mev boost works when using external vc like… (#433)
We are using the ethpandaops collection to run the clients and realized
with a setup using ssv nodes and teku there was no builder-url showing
up in the commands. The mev-boost was only enabled if the validator
client also was enabled which prevents the usage of external vcs. The
pull request fixes that by using the same conditions like you are using
for example for the lighthouse role.

---------

Co-authored-by: Maximilian Stahl <maximilian.stahl@blockscape.network>
Co-authored-by: Rafael Matias <rafael@skyle.net>